gist-ai/requirements.txt
yt-dlp
openai-whisper
ffmpeg
ollama
openai
dotenv
groq
moviepy